Move serde tests, fix deprecation warning
